+++ node/src/integration-test/kotlin/net/corda/node/services/DistributedServiceTests.kt
@@ -11,8 +11,9 @@ import net.corda.core.messaging.StateMachineUpdate
 import net.corda.core.messaging.startFlow
 import net.corda.core.node.NodeInfo
 import net.corda.core.serialization.OpaqueBytes
-import net.corda.flows.CashCommand
-import net.corda.flows.CashFlow
+import net.corda.core.transactions.SignedTransaction
+import net.corda.flows.CashIssueFlow
+import net.corda.flows.CashPaymentFlow
 import net.corda.node.driver.DriverBasedTest
 import net.corda.node.driver.NodeHandle
 import net.corda.node.driver.driver
@@ -35,7 +36,10 @@ class DistributedServiceTests : DriverBasedTest() {
     override fun setup() = driver {
         // Start Alice and 3 notaries in a RAFT cluster
         val clusterSize = 3
-        val testUser = User("test", "test", permissions = setOf(startFlowPermission<CashFlow>()))
+        val testUser = User("test", "test", permissions = setOf(
+                startFlowPermission<CashIssueFlow>(),
+                startFlowPermission<CashPaymentFlow>())
+        )
         val aliceFuture = startNode("Alice", rpcUsers = listOf(testUser))
         val notariesFuture = startNotaryCluster(
                 "Notary",
@@ -135,15 +139,15 @@ class DistributedServiceTests : DriverBasedTest() {
 
     private fun issueCash(amount: Amount<Currency>) {
         val issueHandle = aliceProxy.startFlow(
-                ::CashFlow,
-                CashCommand.IssueCash(amount, OpaqueBytes.of(0), alice.nodeInfo.legalIdentity, raftNotaryIdentity))
+                ::CashIssueFlow,
+                amount, OpaqueBytes.of(0), alice.nodeInfo.legalIdentity, raftNotaryIdentity)
         issueHandle.returnValue.getOrThrow()
     }
 
     private fun paySelf(amount: Amount<Currency>) {
         val payHandle = aliceProxy.startFlow(
-                ::CashFlow,
-                CashCommand.PayCash(amount.issuedBy(alice.nodeInfo.legalIdentity.ref(0)), alice.nodeInfo.legalIdentity))
+                ::CashPaymentFlow,
+                amount.issuedBy(alice.nodeInfo.legalIdentity.ref(0)), alice.nodeInfo.legalIdentity)
         payHandle.returnValue.getOrThrow()
     }
 }