+++ node/src/main/kotlin/net/corda/node/services/statemachine/StateMachineManager.kt
@@ -6,11 +6,13 @@ import co.paralleluniverse.io.serialization.kryo.KryoSerializer
 import co.paralleluniverse.strands.Strand
 import com.codahale.metrics.Gauge
 import com.esotericsoftware.kryo.Kryo
+import com.google.common.collect.HashMultimap
 import com.google.common.util.concurrent.ListenableFuture
 import kotlinx.support.jdk8.collections.removeIf
 import net.corda.core.ThreadBox
 import net.corda.core.bufferUntilSubscribed
 import net.corda.core.crypto.Party
+import net.corda.core.crypto.SecureHash
 import net.corda.core.crypto.commonName
 import net.corda.core.flows.FlowException
 import net.corda.core.flows.FlowLogic
@@ -62,7 +64,7 @@ import javax.annotation.concurrent.ThreadSafe
  * TODO: Timeouts
  * TODO: Surfacing of exceptions via an API and/or management UI
  * TODO: Ability to control checkpointing explicitly, for cases where you know replaying a message can't hurt
- * TODO: Implement stub/skel classes that provide a basic RPC framework on top of this.
+ * TODO: Don't store all active flows in memory, load from the database on demand.
  */
 @ThreadSafe
 class StateMachineManager(val serviceHub: ServiceHubInternal,
@@ -89,15 +91,17 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
 
     // A list of all the state machines being managed by this class. We expose snapshots of it via the stateMachines
     // property.
-    private val mutex = ThreadBox(object {
+    private class InnerState {
         var started = false
         val stateMachines = LinkedHashMap<FlowStateMachineImpl<*>, Checkpoint>()
-        val changesPublisher = PublishSubject.create<Change>()
+        val changesPublisher = PublishSubject.create<Change>()!!
+        val fibersWaitingForLedgerCommit = HashMultimap.create<SecureHash,  FlowStateMachineImpl<*>>()!!
 
         fun notifyChangeObservers(fiber: FlowStateMachineImpl<*>, addOrRemove: AddOrRemove) {
             changesPublisher.bufferUntilDatabaseCommit().onNext(Change(fiber.logic, addOrRemove, fiber.id))
         }
-    })
+    }
+    private val mutex = ThreadBox(InnerState())
 
     // True if we're shutting down, so don't resume anything.
     @Volatile private var stopping = false
@@ -152,7 +156,25 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
 
     fun start() {
         restoreFibersFromCheckpoints()
+        listenToLedgerTransactions()
         serviceHub.networkMapCache.mapServiceRegistered.then(executor) { resumeRestoredFibers() }
+    }
+
+    private fun listenToLedgerTransactions() {
+        // Observe the stream of committed, validated transactions and resume fibers that are waiting for them.
+        serviceHub.storageService.validatedTransactions.updates.subscribe { stx ->
+            val hash = stx.id
+            val fibers: Set<FlowStateMachineImpl<*>> = mutex.locked { fibersWaitingForLedgerCommit.removeAll(hash) }
+            if (fibers.isNotEmpty()) {
+                executor.executeASAP {
+                    for (fiber in fibers) {
+                        fiber.logger.info("Transaction $hash has committed to the ledger, resuming")
+                        fiber.waitingForResponse = null
+                        resumeFiber(fiber)
+                    }
+                }
+            }
+        }
     }
 
     private fun decrementLiveFibers() {
@@ -217,8 +239,23 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
 
     private fun resumeRestoredFiber(fiber: FlowStateMachineImpl<*>) {
         fiber.openSessions.values.forEach { openSessions[it.ourSessionId] = it }
-        if (fiber.openSessions.values.any { it.waitingForResponse }) {
-            fiber.logger.info("Restored, pending on receive")
+        val waitingForResponse = fiber.waitingForResponse
+        if (waitingForResponse != null) {
+            if (waitingForResponse is WaitForLedgerCommit) {
+                val stx = databaseTransaction(database) {
+                    serviceHub.storageService.validatedTransactions.getTransaction(waitingForResponse.hash)
+                }
+                if (stx != null) {
+                    fiber.logger.info("Resuming fiber as tx ${waitingForResponse.hash} has committed.")
+                    fiber.waitingForResponse = null
+                    resumeFiber(fiber)
+                } else {
+                    fiber.logger.info("Restored, pending on ledger commit of ${waitingForResponse.hash}")
+                    mutex.locked { fibersWaitingForLedgerCommit.put(waitingForResponse.hash, fiber) }
+                }
+            } else {
+                fiber.logger.info("Restored, pending on receive")
+            }
         } else {
             resumeFiber(fiber)
         }
@@ -241,15 +278,17 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
     private fun onExistingSessionMessage(message: ExistingSessionMessage, sender: Party) {
         val session = openSessions[message.recipientSessionId]
         if (session != null) {
-            session.fiber.logger.trace { "Received $message on $session" }
+            session.fiber.logger.trace { "Received $message on $session from $sender" }
             if (message is SessionEnd) {
                 openSessions.remove(message.recipientSessionId)
             }
             session.receivedMessages += ReceivedSessionMessage(sender, message)
-            if (session.waitingForResponse) {
-                // We only want to resume once, so immediately reset the flag.
-                session.waitingForResponse = false
+            if (resumeOnMessage(message, session)) {
+                // It's important that we reset here and not after the fiber's resumed, in case we receive another message
+                // before then.
+                session.fiber.waitingForResponse = null
                 updateCheckpoint(session.fiber)
+                session.fiber.logger.debug { "About to resume due to $message" }
                 resumeFiber(session.fiber)
             }
         } else {
@@ -257,7 +296,7 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
             if (peerParty != null) {
                 if (message is SessionConfirm) {
                     logger.debug { "Received session confirmation but associated fiber has already terminated, so sending session end" }
-                    sendSessionMessage(peerParty, SessionEnd(message.initiatedSessionId, null))
+                    sendSessionMessage(peerParty, NormalSessionEnd(message.initiatedSessionId))
                 } else {
                     logger.trace { "Ignoring session end message for already closed session: $message" }
                 }
@@ -265,6 +304,14 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
                 logger.warn("Received a session message for unknown session: $message")
             }
         }
+    }
+
+    // We resume the fiber if it's received a response for which it was waiting for or it's waiting for a ledger
+    // commit but a counterparty flow has ended with an error (in which case our flow also has to end)
+    private fun resumeOnMessage(message: ExistingSessionMessage, session: FlowSession): Boolean {
+        val waitingForResponse = session.fiber.waitingForResponse
+        return (waitingForResponse as? ReceiveRequest<*>)?.session === session ||
+                waitingForResponse is WaitForLedgerCommit && message is ErrorSessionEnd
     }
 
     private fun onSessionInit(sessionInit: SessionInit, sender: Party) {
@@ -345,14 +392,14 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
             processIORequest(ioRequest)
             decrementLiveFibers()
         }
-        fiber.actionOnEnd = { errorResponse: Pair<FlowException, Boolean>? ->
+        fiber.actionOnEnd = { exception, propagated ->
             try {
                 fiber.logic.progressTracker?.currentStep = ProgressTracker.DONE
                 mutex.locked {
                     stateMachines.remove(fiber)?.let { checkpointStorage.removeCheckpoint(it) }
                     notifyChangeObservers(fiber, AddOrRemove.REMOVE)
                 }
-                endAllFiberSessions(fiber, errorResponse)
+                endAllFiberSessions(fiber, exception, propagated)
             } finally {
                 fiber.commitTransaction()
                 decrementLiveFibers()
@@ -367,10 +414,10 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
         }
     }
 
-    private fun endAllFiberSessions(fiber: FlowStateMachineImpl<*>, errorResponse: Pair<FlowException, Boolean>?) {
+    private fun endAllFiberSessions(fiber: FlowStateMachineImpl<*>, exception: Throwable?, propagated: Boolean) {
         openSessions.values.removeIf { session ->
             if (session.fiber == fiber) {
-                session.endSession(errorResponse)
+                session.endSession(exception, propagated)
                 true
             } else {
                 false
@@ -378,22 +425,21 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
         }
     }
 
-    private fun FlowSession.endSession(errorResponse: Pair<FlowException, Boolean>?) {
+    private fun FlowSession.endSession(exception: Throwable?, propagated: Boolean) {
         val initiatedState = state as? Initiated ?: return
-        val propagatedException = errorResponse?.let {
-            val (exception, propagated) = it
-            if (propagated) {
-                // This exception was propagated to us. We only propagate it down the invocation chain to the flow that
-                // initiated us, not to flows we've started sessions with.
-                if (initiatingParty != null) exception else null
+        val sessionEnd = if (exception == null) {
+            NormalSessionEnd(initiatedState.peerSessionId)
+        } else {
+            val errorResponse = if (exception is FlowException && (!propagated || initiatingParty != null)) {
+                // Only propagate this FlowException if our local flow threw it or it was propagated to us and we only
+                // pass it down invocation chain to the flow that initiated us, not to flows we've started sessions with.
+                exception
             } else {
-                exception // Our local flow threw the exception so propagate it
+                null
             }
+            ErrorSessionEnd(initiatedState.peerSessionId, errorResponse)
         }
-        sendSessionMessage(
-                initiatedState.peerParty,
-                SessionEnd(initiatedState.peerSessionId, propagatedException),
-                fiber)
+        sendSessionMessage(initiatedState.peerParty, sessionEnd, fiber)
         recentlyClosedSessions[ourSessionId] = initiatedState.peerParty
     }
 
@@ -424,6 +470,7 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
      * Note that you must be on the [executor] thread.
      */
     fun <T> add(logic: FlowLogic<T>): FlowStateMachine<T> {
+        // TODO: Check that logic has @Suspendable on its call method.
         executor.checkOnThread()
         // We swap out the parent transaction context as using this frequently leads to a deadlock as we wait
         // on the flow completion future inside that context. The problem is that any progress checkpoints are
@@ -457,8 +504,10 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
     private fun resumeFiber(fiber: FlowStateMachineImpl<*>) {
         // Avoid race condition when setting stopping to true and then checking liveFibers
         incrementLiveFibers()
-        if (!stopping) executor.executeASAP {
-            fiber.resume(scheduler)
+        if (!stopping) {
+            executor.executeASAP {
+                fiber.resume(scheduler)
+            }
         } else {
             fiber.logger.debug("Not resuming as SMM is stopping.")
             decrementLiveFibers()
@@ -466,6 +515,7 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
     }
 
     private fun processIORequest(ioRequest: FlowIORequest) {
+        executor.checkOnThread()
         if (ioRequest is SendRequest) {
             if (ioRequest.message is SessionInit) {
                 openSessions[ioRequest.session.ourSessionId] = ioRequest.session
@@ -474,6 +524,24 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
             if (ioRequest !is ReceiveRequest<*>) {
                 // We sent a message, but don't expect a response, so re-enter the continuation to let it keep going.
                 resumeFiber(ioRequest.session.fiber)
+            }
+        } else if (ioRequest is WaitForLedgerCommit) {
+            // Is it already committed?
+            val stx = databaseTransaction(database) {
+                serviceHub.storageService.validatedTransactions.getTransaction(ioRequest.hash)
+            }
+            if (stx != null) {
+                resumeFiber(ioRequest.fiber)
+            } else {
+                // No, then register to wait.
+                //
+                // We assume this code runs on the server thread, which is the only place transactions are committed
+                // currently. When we liberalise our threading somewhat, handing of wait requests will need to be
+                // reworked to make the wait atomic in another way. Otherwise there is a race between checking the
+                // database and updating the waiting list.
+                mutex.locked {
+                    fibersWaitingForLedgerCommit[ioRequest.hash] += ioRequest.fiber
+                }
             }
         }
     }
@@ -514,10 +582,9 @@ class StateMachineManager(val serviceHub: ServiceHubInternal,
             val flow: FlowLogic<*>,
             val ourSessionId: Long,
             val initiatingParty: Party?,
-            var state: FlowSessionState,
-            @Volatile var waitingForResponse: Boolean = false
-    ) {
-        val receivedMessages = ConcurrentLinkedQueue<ReceivedSessionMessage<ExistingSessionMessage>>()
+            var state: FlowSessionState)
+    {
+        val receivedMessages = ConcurrentLinkedQueue<ReceivedSessionMessage<*>>()
         val fiber: FlowStateMachineImpl<*> get() = flow.stateMachine as FlowStateMachineImpl<*>
     }
 }